NOJIRA fjerne kodeliste-tabell-bruk
