+++ behandlingslager/domene/src/main/java/no/nav/foreldrepenger/behandlingslager/TraverseEntityGraphFactory.java
@@ -7,10 +7,23 @@ import no.nav.foreldrepenger.behandlingslager.behandling.søknad.SøknadEntitet;
 import no.nav.foreldrepenger.behandlingslager.diff.TraverseGraph;
 import no.nav.foreldrepenger.behandlingslager.diff.TraverseGraphConfig;
 import no.nav.foreldrepenger.behandlingslager.diff.TraverseJpaEntityGraphConfig;
-import no.nav.foreldrepenger.behandlingslager.kodeverk.Kodeliste;
-import no.nav.foreldrepenger.behandlingslager.kodeverk.KodeverkTabell;
+import no.nav.foreldrepenger.behandlingslager.fagsak.Dekningsgrad;
+import no.nav.foreldrepenger.behandlingslager.kodeverk.Kodeverdi;
+import no.nav.foreldrepenger.behandlingslager.uttak.fp.SamtidigUttaksprosent;
+import no.nav.foreldrepenger.behandlingslager.uttak.fp.Trekkdager;
+import no.nav.foreldrepenger.behandlingslager.uttak.fp.Utbetalingsgrad;
+import no.nav.foreldrepenger.behandlingslager.virksomhet.Arbeidsgiver;
+import no.nav.foreldrepenger.behandlingslager.virksomhet.OrgNummer;
 import no.nav.foreldrepenger.domene.tid.DatoIntervallEntitet;
 import no.nav.foreldrepenger.domene.tid.ÅpenDatoIntervallEntitet;
+import no.nav.foreldrepenger.domene.typer.AktørId;
+import no.nav.foreldrepenger.domene.typer.Beløp;
+import no.nav.foreldrepenger.domene.typer.EksternArbeidsforholdRef;
+import no.nav.foreldrepenger.domene.typer.InternArbeidsforholdRef;
+import no.nav.foreldrepenger.domene.typer.JournalpostId;
+import no.nav.foreldrepenger.domene.typer.PersonIdent;
+import no.nav.foreldrepenger.domene.typer.Saksnummer;
+import no.nav.foreldrepenger.domene.typer.Stillingsprosent;
 
 public final class TraverseEntityGraphFactory {
     private TraverseEntityGraphFactory() {
@@ -30,9 +43,23 @@ public final class TraverseEntityGraphFactory {
         config.addRootClasses(Behandling.class, SøknadEntitet.class);
         config.setInclusionFilter(inclusionFilter);
 
-        config.addLeafClasses(KodeverkTabell.class);
-        config.addLeafClasses(Kodeliste.class);
+        config.addLeafClasses(AktørId.class);
+        config.addLeafClasses(Saksnummer.class);
+        config.addLeafClasses(JournalpostId.class);
+        config.addLeafClasses(PersonIdent.class);
+        config.addLeafClasses(OrgNummer.class);
+        config.addLeafClasses(EksternArbeidsforholdRef.class);
+        config.addLeafClasses(InternArbeidsforholdRef.class);
+        config.addLeafClasses(Stillingsprosent.class);
+        config.addLeafClasses(Arbeidsgiver.class);
+        config.addLeafClasses(Beløp.class);
+        config.addLeafClasses(Utbetalingsgrad.class);
+        config.addLeafClasses(SamtidigUttaksprosent.class);
+        config.addLeafClasses(Trekkdager.class);
+        config.addLeafClasses(Dekningsgrad.class);
+
         config.addLeafClasses(DatoIntervallEntitet.class, ÅpenDatoIntervallEntitet.class);
+        config.addLeafClasses(Kodeverdi.class);
 
         config.addLeafClasses(leafClasses);
         return new TraverseGraph(config);
