--- behandlingslager/domene/src/main/java/no/nav/foreldrepenger/behandlingslager/geografisk/Postnummer.java
@@ -1,93 +1,0 @@
-package no.nav.foreldrepenger.behandlingslager.geografisk;
-
-import java.time.LocalDate;
-import java.util.Objects;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-import no.nav.foreldrepenger.behandlingslager.BaseEntitet;
-
-
-@Entity(name = "Postnummer")
-@Table(name = "POSTSTED")
-public class Postnummer extends BaseEntitet {
-
-    @Id
-    @Column(name = "poststednummer", updatable = false, nullable = false)
-    private String poststednummer;
-
-    @Column(name = "poststednavn", nullable = false)
-    private String poststednavn;
-
-    @Column(name = "gyldigfom", nullable = false)
-    private LocalDate gyldigFom;
-
-    @Column(name = "gyldigtom", nullable = false)
-    private LocalDate gyldigTom;
-
-    public Postnummer() {
-        //For Hibernate
-    }
-
-    public Postnummer(String poststednummer, String poststednavn, LocalDate gyldigFom, LocalDate gyldigTom) {
-        this.poststednummer = poststednummer;
-        this.poststednavn = poststednavn;
-        this.gyldigFom = gyldigFom;
-        this.gyldigTom = gyldigTom;
-    }
-
-    public String getPoststednummer() {
-        return poststednummer;
-    }
-
-    public String getPoststednavn() {
-        return poststednavn;
-    }
-
-    public LocalDate getGyldigFom() {
-        return gyldigFom;
-    }
-
-    public LocalDate getGyldigTom() {
-        return gyldigTom;
-    }
-
-    public void setPoststednavn(String poststednavn) {
-        this.poststednavn = poststednavn;
-    }
-
-    public void setGyldigFom(LocalDate gyldigFom) {
-        this.gyldigFom = gyldigFom;
-    }
-
-    public void setGyldigTom(LocalDate gyldigTom) {
-        this.gyldigTom = gyldigTom;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Postnummer that = (Postnummer) o;
-        return poststednummer.equals(that.poststednummer);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(poststednummer);
-    }
-
-    @Override
-    public String toString() {
-        return "Postnummer{" +
-                "poststednummer='" + poststednummer + '\'' +
-                ", poststednavn='" + poststednavn + '\'' +
-                ", gyldigFom=" + gyldigFom +
-                ", gyldigTom=" + gyldigTom +
-                '}';
-    }
-
-}
