+++ behandlingslager/domene/src/main/java/no/nav/foreldrepenger/behandlingslager/geografisk/Poststed.java
@@ -1,24 +1,93 @@
 package no.nav.foreldrepenger.behandlingslager.geografisk;
 
-import javax.persistence.DiscriminatorValue;
-import javax.persistence.Entity;
+import java.time.LocalDate;
+import java.util.Objects;
 
-import no.nav.foreldrepenger.behandlingslager.kodeverk.Kodeliste;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import no.nav.foreldrepenger.behandlingslager.BaseEntitet;
+
 
 @Entity(name = "Poststed")
-@DiscriminatorValue(Poststed.DISCRIMINATOR)
-public class Poststed extends Kodeliste {
+@Table(name = "POSTSTED")
+public class Poststed extends BaseEntitet {
 
-    public static final String DISCRIMINATOR = "POSTSTED";
-    public static final Poststed OSLO = new Poststed("0103");
+    @Id
+    @Column(name = "poststednummer", updatable = false, nullable = false)
+    private String poststednummer;
 
+    @Column(name = "poststednavn", nullable = false)
+    private String poststednavn;
 
-    Poststed() {
-        // Hibernate trenger den
+    @Column(name = "gyldigfom", nullable = false)
+    private LocalDate gyldigFom;
+
+    @Column(name = "gyldigtom", nullable = false)
+    private LocalDate gyldigTom;
+
+    public Poststed() {
+        //For Hibernate
     }
 
-    public Poststed(String kode) {
-        super(kode, DISCRIMINATOR);
+    public Poststed(String poststednummer, String poststednavn, LocalDate gyldigFom, LocalDate gyldigTom) {
+        this.poststednummer = poststednummer;
+        this.poststednavn = poststednavn;
+        this.gyldigFom = gyldigFom;
+        this.gyldigTom = gyldigTom;
+    }
+
+    public String getPoststednummer() {
+        return poststednummer;
+    }
+
+    public String getPoststednavn() {
+        return poststednavn;
+    }
+
+    public LocalDate getGyldigFom() {
+        return gyldigFom;
+    }
+
+    public LocalDate getGyldigTom() {
+        return gyldigTom;
+    }
+
+    public void setPoststednavn(String poststednavn) {
+        this.poststednavn = poststednavn;
+    }
+
+    public void setGyldigFom(LocalDate gyldigFom) {
+        this.gyldigFom = gyldigFom;
+    }
+
+    public void setGyldigTom(LocalDate gyldigTom) {
+        this.gyldigTom = gyldigTom;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Poststed that = (Poststed) o;
+        return poststednummer.equals(that.poststednummer);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(poststednummer);
+    }
+
+    @Override
+    public String toString() {
+        return "Postnummer{" +
+                "poststednummer='" + poststednummer + '\'' +
+                ", poststednavn='" + poststednavn + '\'' +
+                ", gyldigFom=" + gyldigFom +
+                ", gyldigTom=" + gyldigTom +
+                '}';
     }
 
 }
