+++ behandlingslager/domene/src/main/java/no/nav/foreldrepenger/behandlingslager/geografisk/PoststedKodeverkRepository.java
@@ -1,5 +1,7 @@
 package no.nav.foreldrepenger.behandlingslager.geografisk;
 
+import java.time.LocalDate;
+import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
@@ -8,10 +10,15 @@ import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.persistence.TypedQuery;
 
+import org.hibernate.jpa.QueryHints;
+
 import no.nav.vedtak.felles.jpa.HibernateVerktøy;
+import no.nav.vedtak.konfig.Tid;
 
 @ApplicationScoped
 public class PoststedKodeverkRepository {
+
+    private static final String SYNK_POSTNUMMER = "SYNK";
 
     private EntityManager entityManager;
 
@@ -20,16 +27,37 @@ public class PoststedKodeverkRepository {
     }
 
     @Inject
-    public PoststedKodeverkRepository( EntityManager entityManager) {
+    public PoststedKodeverkRepository(EntityManager entityManager) {
         Objects.requireNonNull(entityManager, "entityManager"); //$NON-NLS-1$
         this.entityManager = entityManager;
     }
 
-    public Optional<Poststed> finnPoststed(String postnummer) {
-        TypedQuery<Poststed> query = entityManager.createQuery("from Poststed where kode=:kode",
-                Poststed.class);
-            query.setParameter("kode", postnummer);
-            return HibernateVerktøy.hentUniktResultat(query);
+    public List<Poststed> hentAllePostnummer() {
+        TypedQuery<Poststed> query = entityManager.createQuery("from Postnummer p where poststednummer <> :postnr", Poststed.class)
+                .setParameter("postnr", SYNK_POSTNUMMER)
+                .setHint(QueryHints.HINT_READONLY, "true");
+        return query.getResultList();
     }
 
+    public Optional<Poststed> finnPoststed(String postnummer) {
+        TypedQuery<Poststed> query = entityManager.createQuery("from Postnummer p where poststednummer = :postnr", Poststed.class)
+                .setParameter("postnr", postnummer)
+                .setHint(QueryHints.HINT_READONLY, "true");
+        return HibernateVerktøy.hentUniktResultat(query);
+    }
+
+    public void lagrePostnummer(Poststed postnummer) {
+        entityManager.persist(postnummer);
+    }
+
+    public LocalDate getPostnummerKodeverksDato() {
+        return finnPoststed(SYNK_POSTNUMMER).map(Poststed::getGyldigFom).orElse(Tid.TIDENES_BEGYNNELSE);
+    }
+
+    public void setPostnummerKodeverksDato(String versjon, LocalDate synkDato) {
+        var postnummer = finnPoststed(SYNK_POSTNUMMER).orElseGet(() -> new Poststed(SYNK_POSTNUMMER, versjon, synkDato, Tid.TIDENES_ENDE));
+        postnummer.setGyldigFom(synkDato);
+        postnummer.setPoststednavn("VERSJON" + versjon);
+        entityManager.persist(postnummer);
+    }
 }
