--- behandlingslager/domene/src/main/java/no/nav/foreldrepenger/behandlingslager/kodeverk/Kodeverk.java
@@ -1,106 +1,0 @@
-package no.nav.foreldrepenger.behandlingslager.kodeverk;
-
-import javax.persistence.Column;
-import javax.persistence.Convert;
-import javax.persistence.Entity;
-import javax.persistence.Table;
-
-import no.nav.vedtak.felles.jpa.converters.BooleanToStringConverter;
-
-/**
- *
- * Representerer et kodeverk brukt i applikasjonen. Hvert kodeverk har en eier (default er 'VL' dvs. det er internt
- * kodeverk)
- *
- * Dersom noen av synk flaggene er satt skal dette kodeverket synkroniseres automatisk med kodeverk eier (fungerer bare
- * hvis eier støtter dette og vi har implementert det).
- */
-@Entity(name = "Kodeverk")
-@Table(name = Kodeverk.KODEVERK)
-public class Kodeverk extends KodeverkTabell {
-
-    static final String KODEVERK = "KODEVERK";
-
-    @Column(name = "kodeverk_eier", nullable = false)
-    private String kodeverkEier;
-
-    @Column(name = "kodeverk_eier_ref")
-    private String kodeverkEierReferanse;
-
-    @Column(name = "kodeverk_eier_ver")
-    private String kodeverkEierVersjon;
-
-    @Column(name = "kodeverk_eier_navn")
-    private String kodeverkEierNavn;
-
-    /** Hvorvidt nye koder skal legges til ved synk. */
-    @Convert(converter = BooleanToStringConverter.class)
-    @Column(name = "kodeverk_synk_nye", nullable = false)
-    private boolean synkNyeKoderFraKodeverEier;
-
-    /** Hvorvidt endringer i gyldig dato (fom, tom) og navn skal oppdateres ved synk. */
-    @Convert(converter = BooleanToStringConverter.class)
-    @Column(name = "kodeverk_synk_eksisterende", nullable = false)
-    private boolean synkEksisterendeKoderFraKodeverkEier;
-
-    @Convert(converter = BooleanToStringConverter.class)
-    @Column(name = "sammensatt")
-    private Boolean sammensatt;
-
-    public Kodeverk() {
-        // proxy for hibernate
-    }
-
-    /**
-     * For å kunne skape instanser hvor man ikke leser opp fra databasen
-     * slik som ved enhetstesting.
-     */
-    public Kodeverk(String kode, String kodeverkEier, String kodeverkEierVersjon, String kodeverkEierNavn,
-                    boolean synkNyeKoderFraKodeverEier, boolean synkEksisterendeKoderFraKodeverkEier, boolean sammensatt){
-        super(kode);
-        this.kodeverkEier = kodeverkEier;
-        this.kodeverkEierVersjon = kodeverkEierVersjon;
-        this.kodeverkEierNavn = kodeverkEierNavn;
-        this.synkNyeKoderFraKodeverEier = synkNyeKoderFraKodeverEier;
-        this.synkEksisterendeKoderFraKodeverkEier = synkEksisterendeKoderFraKodeverkEier;
-        this.sammensatt = sammensatt;
-    }
-
-    public String getKodeverkEier() {
-        return kodeverkEier;
-    }
-
-    public String getKodeverkEierNavn() {
-        return kodeverkEierNavn;
-    }
-    
-    @Override
-    public String getOffisiellKode() {
-        return getKode();
-    }
-    
-    @Override
-    public String getKodeverk() {
-        return KODEVERK;
-    }
-
-    public String getKodeverkEierReferanse() {
-        return kodeverkEierReferanse;
-    }
-
-    public String getKodeverkEierVersjon() {
-        return kodeverkEierVersjon;
-    }
-
-    public boolean getSynkEksisterendeKoderFraKodeverkEier() {
-        return synkEksisterendeKoderFraKodeverkEier;
-    }
-
-    public boolean getSynkNyeKoderFraKodeverEier() {
-        return synkNyeKoderFraKodeverEier;
-    }
-
-    public Boolean getSammensatt() {
-        return sammensatt;
-    }
-}
