--- behandlingslager/domene/src/main/resources/META-INF/pu-default.felles-kodeverk.orm.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<entity-mappings xmlns="http://xmlns.jcp.org/xml/ns/persistence/orm"
-                 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                 xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/persistence/orm
-      http://xmlns.jcp.org/xml/ns/persistence/orm_2_1.xsd"
-                 version="2.1">
-
-    <sequence-generator name="SEQ_KODELISTE" allocation-size="50" sequence-name="SEQ_KODELISTE" />
-
-    <entity class="no.nav.foreldrepenger.behandlingslager.kodeverk.Kodeliste" />
-    <entity class="no.nav.foreldrepenger.behandlingslager.kodeverk.Kodeverk" />
-
-</entity-mappings>
