+++ behandlingslager/domene/src/main/resources/META-INF/pu-default.kodeverk.orm.xml
@@ -5,13 +5,6 @@
       http://xmlns.jcp.org/xml/ns/persistence/orm_2_1.xsd"
     version="2.1">
 
-    <sequence-generator name="SEQ_KODELISTE_RELASJON" allocation-size="50" sequence-name="SEQ_KODELISTE_RELASJON" />
-	<sequence-generator name="SEQ_KODELISTE_NAVN_I18N" allocation-size="50" sequence-name="SEQ_KODELISTE_NAVN_I18N" />
-
-	<!-- legacy DokumentTypeId er som synkes -->
-	<entity class="no.nav.foreldrepenger.behandlingslager.kodeverk.KodelisteNavnI18N" />
-
 	<entity class="no.nav.foreldrepenger.behandlingslager.geografisk.Poststed" />
-	<entity class="no.nav.foreldrepenger.behandlingslager.geografisk.Postnummer" />
 
 </entity-mappings>
