--- infrastrukturtjenester/kodeverk/README.md
@@ -1,18 +1,0 @@
-# Kodeverk
-
-Modulen gir støttetjenester for integrasjon mot NAV's sentrale kodeverkforvaltning. Kodeverk som
-tilbys av kodeverkforvaltningen kan konfigureres til å synkroniseres til vedtaksløsningen.
-
-## Hensikten
-
-* Synkronisere vedtaksløsningens kodeverk slik at de til en hver tid er oppdatert i forhold til endringer 
-i den sentrale kodeverkforvaltning.
-* Integrerer mot webservice for kodeverkforvaltning for å hente oppdaterte kodeverksdata. 
-* Behandler kodeverk både av type enkel kodeliste og sammensatt kodeverk.
-
-## Brukes av
-
-Modulen brukes av Vedtaksløsningen  
-
-## Integrasjoner
-* kodeverk-klient i vl-felles
