--- infrastrukturtjenester/kodeverk/pom.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>infrastrukturtjenester-pom</artifactId>
-        <groupId>no.nav.foreldrepenger</groupId>
-        <version>${revision}${sha1}${changelist}</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>kodeverk</artifactId>
-    <packaging>jar</packaging>
-    <name>FPSAK :: Infrastrukturtjenester - Kodeverk</name>
-
-    <dependencies>
-        <!-- Prosjekt interne avhengigheter -->
-        <dependency>
-            <groupId>no.nav.foreldrepenger</groupId>
-            <artifactId>behandlingslager-domene</artifactId>
-        </dependency>
-        
-        <!-- Infrastrukturtjenester avhengigheter -->
-        <dependency>
-            <groupId>no.nav.foreldrepenger</groupId>
-            <artifactId>batch</artifactId>
-        </dependency>
-
-		<!-- Integrasjontjenester avhengigheter -->
-		<dependency>
-            <groupId>no.nav.foreldrepenger.felles.integrasjon</groupId>
-            <artifactId>kodeverk-klient</artifactId>
-        </dependency>
-		
-        <!-- Test -->
-        <dependency>
-            <groupId>no.nav.foreldrepenger.felles</groupId>
-            <artifactId>felles-testutilities</artifactId>
-            <scope>test</scope>
-        </dependency>
-        
-        <dependency>
-			<groupId>no.nav.foreldrepenger</groupId>
-			<artifactId>migreringer</artifactId>
-			<type>test-jar</type>
-			<scope>test</scope>
-		</dependency>
-    </dependencies>
-</project>
