+++ infrastrukturtjenester/pom.xml
@@ -14,7 +14,6 @@
 
     <modules>
         <module>batch</module>
-        <module>kodeverk</module>
         <module>poststed</module>
     </modules>
