+++ pom.xml
@@ -420,11 +420,6 @@
             </dependency>
             <dependency>
                 <groupId>no.nav.foreldrepenger</groupId>
-                <artifactId>kodeverk</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>no.nav.foreldrepenger</groupId>
                 <artifactId>poststed</artifactId>
                 <version>${project.version}</version>
             </dependency>
