+++ web/src/main/java/no/nav/foreldrepenger/web/app/jackson/JacksonJsonConfig.java
@@ -92,7 +92,6 @@ public class JacksonJsonConfig implements ContextResolver<ObjectMapper> {
     private static void addSerializers(SimpleModule module, boolean serialiserKodelisteNavn) {
         module.addSerializer(new KodeverdiMedNavnSerializer(serialiserKodelisteNavn));
         module.addSerializer(new KodelisteSerializer(serialiserKodelisteNavn));
-        module.addSerializer(new KodeverkSerializer(serialiserKodelisteNavn));
     }
 
     /**
