--- web/src/main/java/no/nav/foreldrepenger/web/app/jackson/KodeverkSerializer.java
@@ -1,43 +1,0 @@
-package no.nav.foreldrepenger.web.app.jackson;
-
-import java.io.IOException;
-
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import com.fasterxml.jackson.databind.ser.std.StdSerializer;
-
-import no.nav.foreldrepenger.behandlingslager.kodeverk.KodeverkTabell;
-
-/**
- * Enkel serialisering av KodeverkTabell klasser, uten at disse trenger @JsonIgnore eller lignende. Deserialisering går
- * av seg selv normalt (får null for andre felter).
- */
-public class KodeverkSerializer extends StdSerializer<KodeverkTabell> {
-
-    /** dropper navn hvis false (trenger da ikke refreshe navn fra db.). Default false */
-    private boolean serialiserKodelisteNavn;
-
-    public KodeverkSerializer() {
-        this(false);
-    }
-
-    public KodeverkSerializer(boolean serialiserKodelisteNavn) {
-        super(KodeverkTabell.class);
-        this.serialiserKodelisteNavn = serialiserKodelisteNavn;
-    }
-
-    @Override
-    public void serialize(KodeverkTabell value, JsonGenerator jgen, SerializerProvider provider) throws IOException {
-
-        jgen.writeStartObject();
-
-        jgen.writeStringField("kode", value.getKode());
-        
-        if (serialiserKodelisteNavn) {
-            jgen.writeStringField("navn", value.getNavn());
-        }
-
-        jgen.writeEndObject();
-    }
-
-}
